requirements.txt
pandas>=1.5.0
numpy>=1.20.0
matplotlib>=3.5.0
seaborn>=0.11.0
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
